feat: Split Module
Split the hilt Module interface into BinderModule and ProviderModule

Resolves: DCMAW-6349
